requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.10.0
astroid==4.0.1
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
dill==0.4.0
Django==5.2.7
django-bulk-update-or-create==0.3.0
et_xmlfile==2.0.0
frozenlist==1.8.0
idna==3.11
invoke==2.2.1
isort==7.0.0
mccabe==0.7.0
multidict==6.7.0
openpyxl==3.1.5
paramiko==4.0.0
platformdirs==4.5.0
propcache==0.4.1
pycparser==2.23
pylint==4.0.2
PyNaCl==1.6.0
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tomli==2.3.0
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
yarl==1.22.0
